+++ src/icons/AvatarIcons/AvatarFour.js
@@ -13,14 +13,14 @@ function AvatarFour({ color }) {
         <g>
           <path
             id="ic_local_airport_24px_8_"
-            fill={ color }
+            fill={color}
             d="M 63.947 67.864 l 1.56 -1.56 l -2.34 -10.14 l 4.289 -4.291 c 0.646 -0.646 0.647 -1.693 0 -2.34 c -0.646 -0.646 -1.693 -0.644 -2.339 0 l -4.291 4.291 l -10.14 -2.338 l -1.56 1.558 l 8.191 4.292 l -4.291 4.289 l -2.731 -0.39 l -1.17 1.17 l 3.512 1.95 l 1.949 3.51 l 1.17 -1.17 l -0.392 -2.729 l 4.292 -4.292 L 63.947 67.864 Z"
           />
         </g>
       </g>
       <circle
         fill="none"
-        stroke={ color }
+        stroke={color}
         strokeMiterlimit="10"
         strokeWidth="2"
         cx="59"
@@ -30,7 +30,7 @@ function AvatarFour({ color }) {
       <g>
         <circle
           fill="none"
-          stroke={ color }
+          stroke={color}
           strokeDasharray="1.9906px, 1.9906px"
           strokeMiterlimit="10"
           strokeWidth="2"
@@ -43,7 +43,7 @@ function AvatarFour({ color }) {
         <g>
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeMiterlimit="10"
             x1="27.353"
             y1="50.392"
@@ -53,7 +53,7 @@ function AvatarFour({ color }) {
 
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeDasharray="1.0154px, 1.0154px"
             strokeMiterlimit="10"
             x1="28.835"
@@ -63,7 +63,7 @@ function AvatarFour({ color }) {
           />
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeMiterlimit="10"
             x1="82.467"
             y1="38.677"
@@ -76,7 +76,7 @@ function AvatarFour({ color }) {
         <g>
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeMiterlimit="10"
             x1="34.229"
             y1="79.6"
@@ -86,7 +86,7 @@ function AvatarFour({ color }) {
 
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeDasharray="1.0154px, 1.0154px"
             strokeMiterlimit="10"
             x1="35.711"
@@ -96,7 +96,7 @@ function AvatarFour({ color }) {
           />
           <line
             fill="none"
-            stroke={ color }
+            stroke={color}
             strokeMiterlimit="10"
             x1="89.343"
             y1="67.885"
@@ -106,119 +106,119 @@ function AvatarFour({ color }) {
         </g>
       </g>
       <path
-        fill={ color }
+        fill={color}
         d="M 39.422 41.245 c 0.318 1.496 -0.638 2.969 -2.135 3.287 l 0 0 c -1.495 0.318 -2.968 -0.639 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.497 0.639 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.97 0.638 3.287 2.135 L 39.422 41.245 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 60.33 36.801 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.968 -0.638 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.637 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 60.33 36.801 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 67.096 35.362 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.496 0.318 -2.969 -0.637 -3.287 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.639 -2.969 2.135 -3.288 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 67.096 35.362 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 74.065 33.881 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.968 -0.637 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.637 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 74.065 33.881 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 82.995 53.345 c 0.152 0.714 -0.304 1.417 -1.02 1.569 l 0 0 c -0.714 0.152 -1.418 -0.305 -1.569 -1.019 l -0.708 -3.329 c -0.152 -0.716 0.306 -1.418 1.02 -1.57 l 0 0 c 0.715 -0.152 1.418 0.303 1.57 1.02 L 82.995 53.345 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 91.053 51.632 c 0.152 0.714 -0.304 1.417 -1.02 1.569 l 0 0 c -0.715 0.152 -1.417 -0.305 -1.569 -1.019 l -0.707 -3.329 c -0.152 -0.716 0.304 -1.417 1.019 -1.57 l 0 0 c 0.716 -0.152 1.418 0.303 1.57 1.02 L 91.053 51.632 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 94.655 50.866 c 0.151 0.714 -0.305 1.417 -1.019 1.569 l 0 0 c -0.715 0.152 -1.419 -0.305 -1.57 -1.019 l -0.708 -3.329 c -0.152 -0.716 0.305 -1.418 1.021 -1.57 l 0 0 c 0.714 -0.152 1.416 0.303 1.568 1.02 L 94.655 50.866 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 84.606 60.158 c 0.151 0.715 -0.307 1.418 -1.021 1.57 l 0 0 c -0.716 0.151 -1.417 -0.305 -1.568 -1.02 L 81.31 57.38 c -0.152 -0.716 0.303 -1.417 1.019 -1.569 l 0 0 c 0.714 -0.152 1.418 0.303 1.57 1.019 L 84.606 60.158 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 86.908 52.513 c 0.151 0.714 -0.305 1.418 -1.02 1.57 l 0 0 c -0.715 0.152 -1.418 -0.305 -1.57 -1.019 l -0.707 -3.329 c -0.152 -0.717 0.305 -1.418 1.019 -1.57 l 0 0 c 0.716 -0.152 1.418 0.303 1.57 1.019 L 86.908 52.513 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 88.596 59.311 c 0.152 0.714 -0.304 1.417 -1.02 1.569 l 0 0 c -0.715 0.151 -1.418 -0.305 -1.569 -1.019 l -0.708 -3.329 c -0.152 -0.716 0.305 -1.417 1.02 -1.57 l 0 0 c 0.716 -0.152 1.418 0.303 1.57 1.02 L 88.596 59.311 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 92.197 58.545 c 0.152 0.714 -0.304 1.417 -1.018 1.569 l 0 0 c -0.716 0.152 -1.419 -0.304 -1.57 -1.019 l -0.708 -3.329 c -0.152 -0.716 0.305 -1.417 1.021 -1.57 l 0 0 c 0.714 -0.151 1.416 0.303 1.568 1.02 L 92.197 58.545 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 95.623 57.604 c 0.152 0.714 -0.304 1.417 -1.019 1.569 l 0 0 c -0.716 0.152 -1.418 -0.304 -1.569 -1.019 l -0.708 -3.329 c -0.152 -0.716 0.304 -1.417 1.02 -1.569 l 0 0 c 0.715 -0.152 1.417 0.303 1.569 1.02 L 95.623 57.604 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 46.636 39.711 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.496 0.318 -2.968 -0.637 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.638 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 46.636 39.711 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 53.402 38.273 c 0.318 1.498 -0.638 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.969 -0.637 -3.287 -2.135 L 46.5 32.457 c -0.318 -1.498 0.638 -2.969 2.135 -3.288 l 0 0 c 1.496 -0.318 2.968 0.637 3.286 2.135 L 53.402 38.273 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 23.328 66.027 c 0.152 0.714 -0.305 1.418 -1.02 1.569 l 0 0 c -0.714 0.151 -1.418 -0.305 -1.569 -1.019 l -0.708 -3.33 c -0.152 -0.716 0.306 -1.418 1.02 -1.569 l 0 0 c 0.715 -0.151 1.418 0.303 1.57 1.02 L 23.328 66.027 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 31.386 64.314 c 0.152 0.715 -0.304 1.418 -1.019 1.569 l 0 0 c -0.716 0.152 -1.418 -0.305 -1.57 -1.019 l -0.707 -3.33 c -0.152 -0.716 0.304 -1.417 1.019 -1.569 l 0 0 c 0.716 -0.152 1.418 0.303 1.57 1.02 L 31.386 64.314 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 34.988 63.549 c 0.151 0.715 -0.305 1.418 -1.019 1.569 l 0 0 c -0.716 0.151 -1.419 -0.305 -1.57 -1.019 l -0.708 -3.33 c -0.152 -0.716 0.305 -1.418 1.021 -1.57 l 0 0 c 0.713 -0.152 1.416 0.303 1.568 1.02 L 34.988 63.549 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 24.939 72.841 c 0.152 0.715 -0.307 1.418 -1.021 1.569 l 0 0 c -0.716 0.152 -1.417 -0.305 -1.569 -1.019 l -0.707 -3.329 c -0.152 -0.717 0.303 -1.418 1.019 -1.569 l 0 0 c 0.714 -0.151 1.419 0.303 1.571 1.019 L 24.939 72.841 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 27.241 65.195 c 0.151 0.715 -0.305 1.418 -1.021 1.57 l 0 0 c -0.713 0.151 -1.417 -0.306 -1.569 -1.02 l -0.707 -3.329 c -0.152 -0.717 0.305 -1.418 1.019 -1.57 l 0 0 c 0.716 -0.152 1.418 0.303 1.57 1.019 L 27.241 65.195 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 28.929 71.993 c 0.151 0.714 -0.304 1.417 -1.02 1.569 l 0 0 c -0.715 0.152 -1.417 -0.305 -1.569 -1.02 l -0.707 -3.328 c -0.152 -0.717 0.304 -1.418 1.019 -1.57 l 0 0 c 0.716 -0.152 1.417 0.303 1.569 1.02 L 28.929 71.993 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 32.531 71.228 c 0.151 0.714 -0.306 1.417 -1.019 1.569 l 0 0 c -0.716 0.151 -1.419 -0.305 -1.571 -1.02 l -0.707 -3.328 c -0.152 -0.717 0.305 -1.418 1.021 -1.57 l 0 0 c 0.713 -0.151 1.416 0.303 1.568 1.02 L 32.531 71.228 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 35.956 70.287 c 0.151 0.714 -0.304 1.417 -1.019 1.568 l 0 0 c -0.716 0.152 -1.418 -0.304 -1.57 -1.019 l -0.707 -3.329 c -0.152 -0.716 0.304 -1.417 1.02 -1.569 l 0 0 c 0.715 -0.152 1.416 0.303 1.568 1.02 L 35.956 70.287 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 45.905 90.984 c 0.318 1.496 -0.638 2.969 -2.135 3.287 l 0 0 c -1.496 0.317 -2.968 -0.639 -3.286 -2.136 l -1.481 -6.969 c -0.318 -1.497 0.638 -2.97 2.134 -3.287 l 0 0 c 1.497 -0.318 2.969 0.638 3.287 2.135 L 45.905 90.984 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 66.813 86.539 c 0.318 1.498 -0.638 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.968 -0.637 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.637 -2.969 2.134 -3.287 l 0 0 c 1.496 -0.318 2.968 0.637 3.286 2.135 L 66.813 86.539 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 73.579 85.102 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.969 -0.637 -3.287 -2.135 l -1.481 -6.97 c -0.318 -1.497 0.638 -2.969 2.135 -3.287 l 0 0 c 1.497 -0.317 2.968 0.638 3.286 2.136 L 73.579 85.102 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 80.548 83.62 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.496 0.317 -2.968 -0.638 -3.286 -2.136 l -1.481 -6.969 c -0.318 -1.498 0.638 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.968 0.638 3.286 2.135 L 80.548 83.62 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 53.119 89.45 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.968 -0.637 -3.286 -2.135 l -1.481 -6.97 c -0.318 -1.497 0.637 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.317 2.968 0.638 3.286 2.135 L 53.119 89.45 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 59.885 88.012 c 0.318 1.498 -0.637 2.969 -2.134 3.287 l 0 0 c -1.497 0.318 -2.969 -0.637 -3.287 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.638 -2.969 2.135 -3.287 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 59.885 88.012 Z"
       />
       <path
-        fill={ color }
+        fill={color}
         d="M 87.396 82.164 c 0.318 1.498 -0.637 2.97 -2.134 3.287 l 0 0 c -1.497 0.318 -2.968 -0.637 -3.286 -2.135 l -1.481 -6.969 c -0.318 -1.498 0.637 -2.969 2.134 -3.287 l 0 0 c 1.497 -0.318 2.968 0.637 3.286 2.135 L 87.396 82.164 Z"
       />
       <g>
