+++ src/icons/ImportIcon.js
@@ -1,0 +1,20 @@
+import React from "react";
+
+function ImportIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 16"
+      width="24"
+      height="16"
+    >
+      <path
+        id="ic_cloud_download_24px"
+        transform="translate(0 -4)"
+        d="M 19.35 10.04 a 7.492 7.492 0 0 0 -14 -2 A 6 6 0 0 0 6 20 H 19 a 4.986 4.986 0 0 0 0.35 -9.96 Z M 17 13 l -5 5 L 7 13 h 3 V 9 h 4 v 4 Z"
+      />
+    </svg>
+  );
+}
+
+export default ImportIcon;
