+++ src/pages/Home/_MapPage.scss
@@ -5,6 +5,7 @@
 @import "./subcomponents/NewUserCity";
 @import "./subcomponents/NewUserSuggestions";
 @import "./subcomponents/FriendCityMap";
+@import "./subcomponents/ImportPopup";
 
 .map-container {
   display: flex;
